requirements.txt
git+https://github.com/faroit/dsdtools.git@WIP_eval
git+https://github.com/faroit/mir_eval.git@bss_eval_images_framewise
pandas
seaborn
matplotlib
numpy
scipy
pytest
cerberus
PyYAML
statsmodels
patsy
